Refactor HowToUseActivity
